Added confirmation buttons for moderation and admin actions

+++ commands/mod-logs/del-mod-log.js
@@ -1,6 +1,6 @@
 const Command = require('../../command-handler/commands/base')
 const { CommandoMessage } = require('../../command-handler/typings')
-const { basicEmbed, docId } = require('../../utils')
+const { basicEmbed, docId, modConfirmation } = require('../../utils')
 const { moderations } = require('../../mongo/schemas')
 const { ModerationSchema } = require('../../mongo/typings')
 const { oneLine } = require('common-tags')
@@ -39,20 +39,23 @@ module.exports = class DelModLogCommand extends Command {
      * @param {string} args.modlogId The mod log id
      */
     async run(message, { modlogId }) {
+        modlogId = modlogId.toLowerCase()
         const { guildId } = message
 
         /** @type {ModerationSchema} */
-        const modLog = await moderations.findOne({ guild: guildId, _id: modlogId.toLowerCase() })
+        const modLog = await moderations.findOne({ guild: guildId, _id: modlogId })
         if (!modLog) {
             return await message.replyEmbed(basicEmbed({
                 color: 'RED', emoji: 'cross', description: 'That id is either invalid or it does not exist.'
             }))
         }
 
+        const confirm = await modConfirmation(message, 'delete modlog', modlogId)
+        if (!confirm) return
         await modLog.deleteOne()
 
         return await message.replyEmbed(basicEmbed({
-            color: 'GREEN', emoji: 'check', description: `Deleted mod log with id \`${modLog._id}\``
+            color: 'GREEN', emoji: 'check', description: `Deleted mod log with id \`${modlogId}\``
         }))
     }
 }
